Update README.md
Color Guessing Game

Welcome to the Color Guessing Game! This is a simple and fun Python-based game where you can test your luck by guessing the correct color from a given list.
Description

In this game, you will start with a balance of ₹1000 and try to guess the correct color chosen randomly by the computer. The game offers a mix of excitement and risk, as guessing correctly will increase your balance, while an incorrect guess will lead to a small penalty. It's an addictive, text-based game that can be played multiple times, with an easy-to-understand interface.

Put your luck to the test, and see how much you can win before you decide to stop!